refactor: split decoration logic from RoomManager and inject managers into Menus
