+++ src/main/java/managers/ClueManager.java
@@ -117,5 +117,4 @@ public class ClueManager {
         }
     }
 
-
 }
